MDL-76147 gradereport_grader: Fix collapsing preferences.
At certain cases I didn't see a tick in view mode. For example
when I click grades only and then switch to totals only a tick
next to total only was not shown. I think this is caused since before
mode was changed in circlurar order i.e.
fullmode->grade only->aggregate only->fullmode. But now any order is
possible so we need to respect it